requirements.txt
send2trash
pillow
numpngw
numpy
matplotlib
imageio
scikit-image
